ml: engines: tflite: Properly ignores warmup time
Every time a new machine learning model is loaded, its first
inference takes longer than usual, due to overall initialization
steps done by the ML accelerators. With that in mind, an extra
inference was being running at the beginning of the application
to avoid mistaken the warmup time with a regular inference time,
but it was always running the inference twice, even after the
warmup was done. This update moves the warmup inference to the
initialization of the interpreter instead of running every
inference twice.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>